removed the function to read the interpolation parameter values from the data from the bulk of pyirf, it is moved to the examples file
