+++ pyirf/interpolation.py
@@ -1,46 +1,8 @@
-"""
-Functions for performing interpolation of IRF to the values read from the data
-"""
+"""Functions for performing interpolation of IRF to the values read from the data"""
 import numpy as np
-import pandas as pd
 import astropy.units as u
 from astropy.io import fits
 from scipy.interpolate import griddata
-
-
-def read_mean_parameters_data(data_file, key, parameters):
-    """
-    Reads a DL2 data fits file and extracts the average values
-    of the parameters for interpolation
-
-    Parameters
-    ----------
-    data_file: ``string``
-        path to the DL2 data file
-    key: ``string``
-        key in the fits file where the parameters are stored
-    parameters: list of ``string``
-        list of parameters as they can be evaluated from DL2 file
-
-    Returns
-    -------
-    interp_pos: tuple
-        tuple of average values of requested parameters
-    """
-
-    # read in the data
-    interp_pos = []  # position for which to interpolate
-    data = pd.read_hdf(data_file, key=key)
-    for par in parameters:
-        # we use here eval function that is considered potentially dangerous
-        # as it can execute arbitrary code, however this is the eval
-        # function from pandas, that is very much limitted to just
-        # the columns read from the file and math operations
-        # so it should be safe here (and it adds a lot of flexibility)
-        val = np.mean(data.eval(par[1]))
-        interp_pos.append(val)
-
-    return tuple(interp_pos)
 
 
 def interpolate_effective_area(aeff_all, pars_all, interp_pars, min_effective_area=1. * u.Unit('m2'), method='linear'):
